move instances and layers to new file
