+++ python/tgn.py
@@ -13,15 +13,15 @@ from mage.tgn.constants import (
     MemoryUpdaterType,
 )
 from mage.tgn.definitions.tgn import (
-    TGN,
-    TGNGraphSumEdgeSelfSupervised,
-    TGNGraphSumSupervised,
-    TGNGraphAttentionSupervised,
+    TGN
 )
 from dataclasses import dataclass
 
-from mage.tgn.definitions.tgn import TGNGraphAttentionEdgeSelfSupervised
+
 from mage.tgn.helper.simple_mlp import MLP
+
+from mage.tgn.definitions.instances import TGNGraphSumEdgeSelfSupervised, TGNGraphSumSupervised, \
+    TGNGraphAttentionSupervised, TGNGraphAttentionEdgeSelfSupervised
 
 
 ###################
